more refactor/more tests
